Fixing Buts at the Close Btn
Tested.. fix it
Renamed some ids in the Bug Command
